Add files via upload
This is an ongoing development, and all code is subject to change. Tirth D Patel has full rights to all the code, including the functions. The image used is copyrighted by D214 / Prospect High School, so please be respectful and do not violate their copyright rights or the rights of the image holders.